feat: controller와 restController 구분하여 재구성
